liblink, cmd/gc: resolve several shift warnings
Address several warnings generated by clang -fsanitize=undefined

R=golang-dev, iant
CC=golang-dev
https://golang.org/cl/43050043

+++ src/liblink/pcln.c
@@ -313,7 +313,7 @@ getvarint(uchar **pp)
 	v = 0;
 	p = *pp;
 	for(shift = 0;; shift += 7) {
-		v |= (*p & 0x7F) << shift;
+		v |= (uint32)(*p & 0x7F) << shift;
 		if(!(*p++ & 0x80))
 			break;
 	}
